+++ components/Changelog/Header/ChangelogHeader.tsx
@@ -1,7 +1,8 @@
 'use client'
 import { useEffect, useState } from 'react'
 import Styles from './styles.module.css'
-import { AppModal as Modal, useDisclosure } from '@/components/ui/Modal'
+import { AppModal as Modal } from '@/components/ui/Modal'
+import { useDisclosure } from '@/hooks/useDisclosure'
 import { DeploymentType, DeploymentTypeColors, DeploymentTypeLabels } from '@/utils/strapi'
 import { useRouter, useSearchParams } from 'next/navigation'
 import { Check, Loader2 } from 'lucide-react'
